Add EventDialogErrorStrings and move keys to ui file
