--- app/lib/timetable/timetable_add_event/src/timetable_add_event_dialog_keys.dart
@@ -1,15 +1,0 @@
-import 'package:flutter/foundation.dart';
-
-@visibleForTesting
-class EventDialogKeys {
-  static const Key saveButton = Key("save-button");
-  static const Key titleTextField = Key("title-field");
-  static const Key courseTile = Key("course-tile");
-  static const Key descriptionTextField = Key("description-field");
-  static const Key startDateField = Key("start-date-field");
-  static const Key startTimeField = Key("start-time-field");
-  static const Key endTimeField = Key("end-time-field");
-  static const Key locationField = Key("location-field");
-  static const Key notifyCourseMembersSwitch =
-      Key("notify-course-members-switch");
-}
