+++ app/lib/timetable/timetable_add_event/src/timetable_add_event_dialog_src.dart
@@ -1,3 +1,2 @@
 export 'timetable_add_event_dialog_controller.dart';
-export 'timetable_add_event_dialog_keys.dart';
 export 'timetable_add_event_event_dialog_api.dart';
