+++ app/lib/timetable/timetable_add_event/timetable_add_event_dialog.dart
@@ -254,7 +254,7 @@ class _TitleField extends StatelessWidget {
               ? 'Titel (z.B. Statistik-Klausur)'
               : 'Titel eingeben (z.B. Sportfest)',
           errorText: controller.showEmptyTitleError
-              ? 'Der Titel darf nicht leer sein.'
+              ? EventDialogErrorStrings.emptyTitle
               : null),
     );
   }
@@ -673,4 +673,21 @@ class _SendNotificationBase extends StatelessWidget {
       description: description != null ? Text(description!) : null,
     );
   }
+}
+
+class EventDialogKeys {
+  static const Key saveButton = Key("save-button");
+  static const Key titleTextField = Key("title-field");
+  static const Key courseTile = Key("course-tile");
+  static const Key descriptionTextField = Key("description-field");
+  static const Key startDateField = Key("start-date-field");
+  static const Key startTimeField = Key("start-time-field");
+  static const Key endTimeField = Key("end-time-field");
+  static const Key locationField = Key("location-field");
+  static const Key notifyCourseMembersSwitch =
+      Key("notify-course-members-switch");
+}
+
+class EventDialogErrorStrings {
+  static const emptyTitle = "Bitte gib einen Titel ein.";
 }
